+++ odoo/local-src/stock_picking_backorder/models/res_partner.py
@@ -12,6 +12,12 @@ class ResPartner(models.Model):
     is_sale_back_order_accepted = fields.Boolean(
         string='Sale backorder accepted',
         default=True,
+        help="Do not let customer order products not in stock"
+    )
+    is_sale_back_order_cancel = fields.Boolean(
+        string='Sale backorder auto-cancel',
+        help=("Automaticaly cancel products that could not be delivered. "
+              "Must be set on delivery addresss")
     )
     is_purchase_back_order_accepted = fields.Boolean(
         string='Purchase backorder accepted',